North pointer now appears conditionally
