+++ GIFrameworkMaps.Web/Views/Map/Partials/PrintPanel.cshtml
@@ -56,7 +56,7 @@
             </div>
         </div>
 
-        <div class="form-check mb-3">
+        <div class="form-check mb-3" id="gifw-print-north-pointer-container">
             <input class="form-check-input" type="checkbox" value="" id="gifw-print-north-pointer">
             <label class="form-check-label" for="gifw-print-north-pointer">Include north pointer?</label>
         </div>
